Sum subroutine project
Implement a basic SBR with parameters and results:
Create a program with a table T of N 32-bit integers (an arbitrary constant N defined with EQU) whose elements we want to sum. The sum will be performed using a sum SBR with two parameters and one result (all on the stack). The parameters will be the starting address of the table (passed by reference) and the number of elements in it (passed by value). The SBR will return the sum of the table's elements as the result. The main program must call the sum SBR and store the sum of the elements in the variable stotal in the data area.