Car in a road project
Design, code, and implement a game consisting of driving a car along a road like the one in the previous figure. The edges of the road are marked with the character ‘#’ (ASCII 35). Initially, the road is straight (columns 8 and 16 of the entire screen) and scrolls from top to bottom. The initial movement rate is 1 row every 0.08 seconds according to the simulation clock. The car (character ‘H’, ASCII 72) is initially in the center of the last row of the screen, and the user moves it across the screen using the keys ‘J’ (ASCII 74 or 106) for left, ‘L’ (ASCII 76 or 108) for right, ‘I’ (ASCII 73 or 105) for up, and ‘K’ (ASCII 75 or 107) for down. If the car crosses the edges of the road, the game ends. Once the game has started, the road will randomly change direction (left, right, or continue straight).

To improve the game, add the key ‘Q’ (ASCII 81 or 113) to end the program, the key ‘+’ (ASCII 43) to double the movement speed (maximum of 1 movement every 0.01 seconds), and the key ‘‐‘ (ASCII 45) to halve the movement speed (minimum of 1 movement every 1.28 seconds).

The keyboard and timer management should be handled with interrupt synchronization and programmed transfer. Any other aspect not specified in the documentation is left to the student's choice.

As an extra, make the road thinner or more sinuous